refactor: move paths helper file into the project config directory
